JS code fixes and refactoring
