+++ js/api/posts/getPosts.mjs
@@ -1,0 +1,11 @@
+import { API_SOCIAL_URL } from "../constants.mjs";
+import { authFetch } from "../authFetch.mjs";
+
+const action = "/posts";
+
+export async function getPosts() {
+  const getPostsURL = `${API_SOCIAL_URL}${action}?_author=true`;
+  const response = await authFetch(getPostsURL);
+
+  return await response.json();
+}
