+++ js/api/posts/index.mjs
@@ -1,4 +1,5 @@
 export * from "./create.mjs";
-export * from "./read.mjs";
+export * from "./getPost.mjs";
+export * from "./getPosts.mjs";
 export * from "./update.mjs";
 export * from "./delete.mjs";
