+++ js/components/showLoader.mjs
@@ -1,4 +1,4 @@
 export function showLoadingIndicator() {
-  const loadingIndicator = document.querySelector(".loading-indicator"); //
+  const loadingIndicator = document.querySelector(".loading-indicator");
   if (loadingIndicator) loadingIndicator.style.display = "block";
 }
