+++ js/index.mjs
@@ -1,4 +1,5 @@
 import "https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js";
+
 import { router } from "./router.mjs";
 
 await router();
